[Part 1/4] Complete. Response body ok
The goal for this part was to implement a GET /seats endpoint to return the available seats and the total rows and columns.
I had to learn how to use set the logging level to debug and read it in order to find out what was causing problems when I tried to insert data into the database.
Additionally, I had to search how to properly use the ResponseEntity class to manipulate the response body.